Lookup category by description in label_tag_mapping.rb
This is for all entities option and it pertains field "Category"of
form in UI for label tag mapping definition.

for options different from all entities, field "Category" was used
for Classification#description and this change ensure that
there will be used Classification#description for category lookup
in that form with option for all entities.